small refactors for handling concurrent simulations
