+++ pkg/jobs/jobs_test.go
@@ -1,1 +1,1 @@
-package bash_test
+package jobs
